- Fixed ambiguity in call to log(size_t) -- now ensure call to log(double) in bic.h
- Removed MPI dependence where it wasn't supposed to be.
  - Fixed makefiles to account for this.
  - New BGL frontend sample config file that builds w/o MPI

+++ bic.h
@@ -79,7 +79,7 @@ namespace cluster {
   ///
   template <typename D>
   double bic(const partition& p, D distance, size_t M) {
-    size_t R = p.size();
+    double R = p.size();
     size_t k = p.num_clusters();
 
     // calculate variance.
